+++ modules/host/variables.tf
@@ -755,7 +755,8 @@ variable "vpc_service_controls" {
         })))
       })))
     })), [])
-    enable_dry_run = optional(bool, false)
+    enable_dry_run             = optional(bool, false)
+    enable_deletion_protection = optional(bool, true)
   }))
   default = {}
 }